+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `PipeLoadCollectionPythonDocs`, parameter `module_names` is now required.
 * More private variables in classes are now marked with `_`.
 * `pipes` parameter to `Config` has now changed to `pipes_and_groups_of_pipes`.
+* To set context on `ProcessCurrentInfo`, now use the `set_context` function.
 
 ### Removed